==> 3_US_Aurora_Registration.ipynb <==
import pandas as pd
import os
from sksurgerynditracker.nditracker import NDITracker
import numpy as np

"""
    Initialize the NDI Aurora EM tracker (Attached to Ultrasound Probe)
"""
SETTINGS_US = {
    "tracker type": "aurora",
    "port": "COM3"
}
tracker_US = NDITracker(SETTINGS_US)
tracker_US.start_tracking()

# {CP} to {US}, Defined and obtained by Calibration Plate and Ultrasound Probe .stl files
CP_to_US_calibration = np.array(
    [
        [1, 0, 0, 90],
        [0, 1, 0, 90],
        [0, 0, 1, 5.48],
        [0, 0, 0, 1]
    ]
)

# {FG} to {EM1}, EM1 is tracked by Aurora Field Generator
_, _, _, FG_to_EM1_calibration, _ = tracker_US.get_frame()

# {EM1} to {FG}
EM1_to_FG_calibration = np.linalg.inv(FG_to_EM1_calibration)

# {FG} to {CP}, calculated by Calibration_Plate_Registration.ipynb
FG_to_CP_calibration = np.load("FG_to_CP_calibration.npy")

# {FG} to {US}
FG_to_US_calibration = np.dot(FG_to_CP_calibration, CP_to_US_calibration)

# {EM1} to {US}
EM1_to_US = np.dot(EM1_to_FG_calibration, FG_to_US_calibration)

# Save EM1_to_US
np.save("EM1_to_US.npy", EM1_to_US)
